requirements.txt
pypdf>=4
sqlalchemy>=2
pgvector>=0.2.5
psycopg[binary,pool]>=3.2
tiktoken
python-dotenv
langchain-text-splitters>=0.2.4
openai>=1.50.0
tabulate
python-docx
openpyxl
python-pptx
readability-lxml
markdownify
requests
ollama
langdetect
google-generativeai>=0.8.0
